Move SettingsCalendarSync* into calendarintegration package
